Add files via upload
This code is used to read and clean the raw JSON files from the UCSD Google Local Dataset. It removes identifiable names of users, and filters the dataset according to the business category from the associated metadata JSON file. This code can run on data for one or multiple U.S. states, just adjust the us_states list as needed.